Split AdaptiveThreePane to AapdativeThreePaneUtils & AdaptiveThreePanePreview
