+++ samples-lib/src/main/java/com/esri/arcgismaps/sample/sampleslib/components/adaptive/AdaptiveThreePane.kt
@@ -14,51 +14,32 @@
  *
  */
 
-package com.esri.arcgismaps.sample.sampleslib.components
+package com.esri.arcgismaps.sample.sampleslib.components.adaptive
 
 import android.content.res.Configuration
 import androidx.activity.compose.BackHandler
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.animateContentSize
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.spring
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.interaction.collectIsDraggedAsState
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.DragIndicator
 import androidx.compose.material.icons.filled.Layers
-import androidx.compose.material3.Button
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.LocalMinimumInteractiveComponentSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.Text
 import androidx.compose.material3.VerticalDragHandle
 import androidx.compose.material3.adaptive.ExperimentalMaterial3AdaptiveApi
 import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
@@ -76,31 +57,24 @@ import androidx.compose.material3.adaptive.navigation.rememberSupportingPaneScaf
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
+import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.hapticfeedback.HapticFeedbackType
-import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
-import androidx.compose.runtime.snapshotFlow
-import com.arcgismaps.mapping.ArcGISScene
-import com.arcgismaps.toolkit.geoviewcompose.SceneView
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.launch
-import kotlin.math.roundToInt
 
 /**
  * Reusable adaptive scaffold for samples that need three layers of UI:
@@ -413,243 +387,6 @@ fun AdaptiveThreePane(
     }
 }
 
-@Composable
-private fun PreviewMainPanePlaceholder(
-    isSupportingPaneVisible: Boolean,
-    isFloatingPaneVisible: Boolean,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.surfaceVariant)
-            .padding(12.dp),
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-                .clip(RoundedCornerShape(12.dp))
-                .background(MaterialTheme.colorScheme.surface.copy(alpha = 0.55f))
-                .border(
-                    width = 1.dp,
-                    color = MaterialTheme.colorScheme.outlineVariant,
-                    shape = RoundedCornerShape(12.dp)
-                ),
-        ) {
-            Column(
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .fillMaxWidth()
-                    .padding(12.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = "Main pane: GeoView",
-                    style = MaterialTheme.typography.headlineMedium,
-                    color = MaterialTheme.colorScheme.onSurface
-                )
-                Text(
-                    text = "Primary content area",
-                    style = MaterialTheme.typography.headlineSmall,
-                    color = MaterialTheme.colorScheme.onSurface
-                )
-                Text(
-                    text = "Supporting visible: $isSupportingPaneVisible",
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-                Text(
-                    text = "Floating visible: $isFloatingPaneVisible",
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-    }
-}
-
-/**
- * Simple heading row for the supporting pane.
- *
- * Keeps title and close affordance consistent across samples.
- *
- * @param title Header text shown at the top of the supporting pane.
- * @param onClose Invoked when the close icon is tapped.
- */
-@Composable
-private fun SupportingPaneHeader(
-    title: String,
-    onClose: () -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(start = 16.dp, end = 4.dp, top = 4.dp, bottom = 4.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.titleMedium,
-            color = MaterialTheme.colorScheme.onSurface,
-        )
-        IconButton(onClick = onClose) {
-            Icon(
-                imageVector = Icons.Default.Close,
-                contentDescription = "Close $title",
-            )
-        }
-    }
-}
-
-/**
- * Floating overlay container that can be dragged within the main pane bounds.
- *
- * Position is maintained as pixel offsets and clamped to remain fully visible as
- * parent or content sizes change.
- *
- * @param containerWidth Available width of the host area in pixels.
- * @param containerHeight Available height of the host area in pixels.
- * @param config Floating behavior and initial position fractions.
- * @param title Header text displayed on the floating card.
- * @param onDismiss Invoked when the floating card dismiss button is tapped.
- * @param content Slot for floating controls/content.
- */
-@Composable
-private fun DraggableFloatingContainer(
-    containerWidth: Int,
-    containerHeight: Int,
-    config: FloatingPaneConfig,
-    title: String,
-    onDismiss: () -> Unit,
-    content: @Composable ColumnScope.() -> Unit,
-) {
-    var rawOffsetX by remember { mutableFloatStateOf(containerWidth * config.initialXFraction) }
-    var rawOffsetY by remember { mutableFloatStateOf(containerHeight * config.initialYFraction) }
-
-    val offsetX by animateFloatAsState(
-        targetValue = rawOffsetX,
-        animationSpec = spring(stiffness = Spring.StiffnessMediumLow),
-        label = "floatingPaneOffsetX",
-    )
-    val offsetY by animateFloatAsState(
-        targetValue = rawOffsetY,
-        animationSpec = spring(stiffness = Spring.StiffnessMediumLow),
-        label = "floatingPaneOffsetY",
-    )
-
-    var paneWidth by remember { mutableIntStateOf(0) }
-    var paneHeight by remember { mutableIntStateOf(0) }
-
-    LaunchedEffect(containerWidth, containerHeight) {
-        // Keep the widget in bounds after rotations/resizes.
-        rawOffsetX =
-            rawOffsetX.coerceIn(0f, (containerWidth - paneWidth).toFloat().coerceAtLeast(0f))
-        rawOffsetY =
-            rawOffsetY.coerceIn(0f, (containerHeight - paneHeight).toFloat().coerceAtLeast(0f))
-    }
-
-    Box(
-        modifier = Modifier
-            .offset { IntOffset(offsetX.roundToInt(), offsetY.roundToInt()) }
-            .onSizeChanged { paneWidth = it.width; paneHeight = it.height }
-            .then(
-                if (config.draggable) {
-                    Modifier.pointerInput(containerWidth, containerHeight) {
-                        detectDragGestures(
-                            onDrag = { change, dragAmount ->
-                                change.consume()
-                                // Clamp dragging so the card cannot be moved off-screen.
-                                val maxX = (containerWidth - paneWidth).toFloat().coerceAtLeast(0f)
-                                val maxY =
-                                    (containerHeight - paneHeight).toFloat().coerceAtLeast(0f)
-                                rawOffsetX = (rawOffsetX + dragAmount.x).coerceIn(0f, maxX)
-                                rawOffsetY = (rawOffsetY + dragAmount.y).coerceIn(0f, maxY)
-                            },
-                        )
-                    }
-                } else Modifier
-            ),
-    ) {
-        FloatingWidgetCard(
-            title = title,
-            onDismiss = onDismiss,
-            content = content,
-        )
-    }
-}
-
-/**
- * Visual card shell for floating controls.
- *
- * This is intentionally style-forward for samples: soft background,
- * compact header, and a clear dismiss action.
- *
- * @param title Header text shown in the floating widget card.
- * @param onDismiss Callback for the dismiss icon.
- * @param modifier Optional modifier for card positioning/styling by callers.
- * @param content Slot for custom controls rendered below the header.
- */
-@Composable
-private fun FloatingWidgetCard(
-    title: String,
-    onDismiss: () -> Unit,
-    modifier: Modifier = Modifier,
-    content: @Composable ColumnScope.() -> Unit,
-) {
-    Box(
-        modifier = modifier
-            .widthIn(max = 320.dp)
-            .clip(RoundedCornerShape(16.dp))
-            .animateContentSize()
-    ) {
-        Box(
-            modifier = Modifier
-                .matchParentSize()
-                .background(MaterialTheme.colorScheme.surfaceContainerHigh.copy(alpha = 0.97f))
-        )
-
-        Column(
-            modifier = Modifier.padding(8.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(6.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.DragIndicator,
-                        contentDescription = null,
-                        tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                        modifier = Modifier.size(24.dp),
-                    )
-                    Text(
-                        text = title,
-                        style = MaterialTheme.typography.titleMedium,
-                        color = MaterialTheme.colorScheme.onSurface,
-                    )
-                }
-                IconButton(
-                    onClick = onDismiss,
-                    modifier = Modifier.size(36.dp),
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Close,
-                        contentDescription = "Dismiss $title",
-                        modifier = Modifier.size(24.dp),
-                    )
-                }
-            }
-            content()
-        }
-    }
-}
-
 // Default split anchor used when the user has no prior pane size preference.
 private const val DEFAULT_EXPANSION_ANCHOR_INDEX = 2
 
@@ -658,114 +395,6 @@ private const val AUTO_CLOSE_EXPANSION_ANCHOR_INDEX = 3
 
 // Auto-close threshold when supporting pane shrinks to a thin strip.
 private const val WIDE_CLOSE_PROPORTION_THRESHOLD = 0.80f
-
-/**
- * User-facing knobs for the adaptive three-pane.
- *
- * @property supportingPaneInitiallyOpen Whether the supporting pane starts open on first composition.
- * @property floatingPaneInitiallyVisible Whether the floating pane is visible on first composition.
- * @property floatingPane Configuration for initial position and drag behavior of the floating pane.
- * @property compactSupportingPaneHeightRatio Height ratio used for supporting pane when compact-stacked.
- */
-data class ThreePaneConfig(
-    val supportingPaneInitiallyOpen: Boolean = true,
-    val floatingPaneInitiallyVisible: Boolean = false,
-    val floatingPane: FloatingPaneConfig = FloatingPaneConfig(),
-    val compactSupportingPaneHeightRatio: Float = 0.5f,
-)
-
-/**
- * Positioning and interaction options for the floating pane.
- *
- * @property initialXFraction Initial x offset as a fraction of the host width.
- * @property initialYFraction Initial y offset as a fraction of the host height.
- * @property draggable Whether dragging is enabled for the floating pane.
- */
-data class FloatingPaneConfig(
-    val initialXFraction: Float = 0.05f,
-    val initialYFraction: Float = 0.05f,
-    val draggable: Boolean = true,
-)
-
-@Composable
-private fun AdaptiveThreePanePreviewContent(config: ThreePaneConfig = ThreePaneConfig()) {
-    SamplePreviewSurface {
-        AdaptiveThreePane(
-            modifier = Modifier.fillMaxSize(),
-            config = config,
-            supportingPaneTitle = "Supporting pane title",
-            floatingPaneTitle = "Floating pane title",
-            mainPane = { _, _ ->
-                // Shows a placeholder GeoView preview instead
-                SceneView(
-                    modifier = Modifier.fillMaxSize(),
-                    arcGISScene = ArcGISScene(),
-                )
-            },
-            supportingPane = { isFloatingPaneVisible, toggleFloatingPane ->
-                PreviewCard {
-                    Text(
-                        text = "M3 UI components like Sliders, Segmented buttons, DropDown controls live here.",
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                    Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Center) {
-                        if (isFloatingPaneVisible) {
-                            Button(onClick = toggleFloatingPane) {
-                                Text("Close floating pane")
-                            }
-                        } else {
-                            OutlinedButton(onClick = toggleFloatingPane) {
-                                Text("Show floating pane")
-                            }
-                        }
-                    }
-                }
-            },
-            floatingPane = {
-                Column(
-                    modifier = Modifier.padding(12.dp),
-                    verticalArrangement = Arrangement.spacedBy(12.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    Text(
-                        text = "Quick action shortcuts:",
-                        style = MaterialTheme.typography.titleMedium,
-                        color = MaterialTheme.colorScheme.onSurface
-                    )
-                    Text(text = "• Reset viewpoint")
-                    Text(text = "• Toggle labels")
-                }
-            },
-        )
-    }
-}
-
-@Composable
-private fun PreviewCard(
-    title: String = "GeoView/Sample controls:",
-    content: @Composable ColumnScope.() -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(12.dp))
-            .background(MaterialTheme.colorScheme.surface)
-            .border(
-                width = 1.dp,
-                color = MaterialTheme.colorScheme.outlineVariant,
-                shape = RoundedCornerShape(12.dp),
-            )
-            .padding(12.dp),
-        verticalArrangement = Arrangement.spacedBy(12.dp)
-    ) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.titleMedium,
-            color = MaterialTheme.colorScheme.onSurface
-        )
-        content()
-    }
-}
 
 @Preview(
     name = "Pane preview - main only day",
@@ -902,4 +531,3 @@ private fun AdaptiveThreePaneCompactRatioPreview() {
         )
     )
 }
-
